+++ pyuvvis/scripts/gwu_spec.py
@@ -1,37 +1,47 @@
-import os, time
+import os, time, shutil, sys
 from optparse import OptionParser
 
-## For testing
-import matplotlib.pyplot as plt, sys
-import matplotlib.cm as cm
-from mpl_toolkits.mplot3d import Axes3D
+import matplotlib.pyplot as plt
 
-from scipy import integrate
-import numpy as np
+### GWU run parameters file ###
+import specparms as sp  #if change filename, change shutil call
 
 
-### ONCE DONE, DELETE AND THEN REPLACE PYUVVIS EVERWHERE
-sys.path.append('../../pyuvvis')
-from pyplots.advanced_plots import spec_surface3d, surf3d, spec_poly3d
-from pandas_utils.dataframeserial import df_load, df_dump
-from pandas_utils.df_attrhandler import transfer_attr
-from core.spec_labeltools import datetime_convert, spectral_convert
-from core.spec_utilities import boxcar, wavelength_slices, divby
-from pyplots.spec_aesthetics import specplot, timeplot, absplot, range_timeplot, _df_colormapper
-from IO.gwu_interfaces import from_timefile_datafile, get_files_in_dir, from_spec_files
-
+### LOCAL IMPORTS
+#sys.path.append('../../pyuvvis')
+#from pyplots.advanced_plots import spec_surface3d, surf3d, spec_poly3d, plot2d
+#from pandas_utils.dataframeserial import df_load, df_dump
+#from pandas_utils.df_attrhandler import transfer_attr
+#from core.spec_labeltools import datetime_convert, spectral_convert, spec_slice
+#from core.spec_utilities import boxcar, wavelength_slices, divby
+#from core.baseline import dynamic_baseline
+#from pyplots.basic_plots import specplot, timeplot, absplot, range_timeplot
+#from pyplots.plot_utils import _df_colormapper, cmget
+#from IO.gwu_interfaces import from_timefile_datafile, get_files_in_dir, from_spec_files
+#from core.baseline import dynamic_baseline
+#from imk_utils import make_root_dir, get_files_in_dir, get_shortname
+#from corr2d import corr_analysis, make_ref, sync_3d, async_3d
 
 ### UPDATE PACKAGE THEN CHANGE THIS
-#from pyuvvis.core.imk_utils import make
-from imk_utils import make_root_dir, get_files_in_dir, get_shortname
+from pyuvvis.pyplots.advanced_plots import spec_surface3d, surf3d, spec_poly3d, plot2d
+from pyuvvis.pandas_utils.dataframeserial import df_load, df_dump
+from pyuvvis.pandas_utils.df_attrhandler import transfer_attr
+from pyuvvis.core.spec_labeltools import datetime_convert, spectral_convert, spec_slice
+from pyuvvis.core.spec_utilities import boxcar, wavelength_slices, divby
+from pyuvvis.core.baseline import dynamic_baseline
+from pyuvvis.pyplots.basic_plots import specplot, timeplot, absplot, range_timeplot
+from pyplots.plot_utils import _df_colormapper, cmget
+from pyuvvis.IO.gwu_interfaces import from_timefile_datafile, get_files_in_dir, from_spec_files
+from pyuvvis.core.baseline import dynamic_baseline
+from pyuvvis.imk_utils import make_root_dir, get_files_in_dir, get_shortname
+from pyuvvis.corr2d import corr_analysis, make_ref, sync_3d, async_3d
 
-def plt_clrsave(outpath, outname):
-    ''' Save plot then clear afterwards '''
-    plt.savefig('%s/%s' %(outpath, outname) )
+
+def plt_clrsave(outpath, outname): # dpi=600):
+    ''' Save plot then clear afterwards.  dpi=600 saves a nicer quality picture for publication.'''
+    plt.savefig('%s/%s' %(outpath, outname)) #, dpi=600 )
     plt.clf()
     
-
-
 if __name__=='__main__':  
     
     img_ignore=['png', 'jpeg', 'tif', 'bmp'] #no .
@@ -48,14 +58,14 @@ if __name__=='__main__':
     parser.add_option('-o', '--outdir',
                       action='store',
                       dest='od',
-                      default='./junk',  #Remove
+                      default='./output',  #Remove
                       help='Path to root directory where output folder is built',) 
 
     parser.add_option('-r', '--runname',
                       action='store',
                       dest='rname',
                       default='', #None causes errors
-                      help='Title of trial/run, used in plot titles and other places.',)     
+                      help='Title of trial/run, used in filenames and other places.',)     
     
     parser.add_option('-i', '--ignore', action='store_true', default=True, 
                       dest='ignore',  help='Ignore the most commonly found image types \
@@ -83,6 +93,9 @@ if __name__=='__main__':
     ### Open logfile to track various IO steps
     start=time.time()
     lf=open(outroot + '/'+ logfilename, 'w')
+    
+    ### Copy of runparameters file for later reference
+    shutil.copyfile('specparms.py', outroot+'/run_parameters')
 
     ### Walk a single directory down    
     walker=os.walk(inroot, topdown=True, onerror=None, followlinks=False)
@@ -102,7 +115,7 @@ if __name__=='__main__':
         try:
             outdir=make_root_dir(od, overwrite=options.clean)  ## BE VERY CAREFUL WITH THIS
         except IOError:
-            raise IOError('Output directory %s already exists.\n  To overwrite, use option -c True.  Use with caution\n\
+            raise IOError('Output directory %s already exists.\n\n  To overwrite, use option -c True.  Use with caution\n\n\
             preexisting data will be deleted from directory!'%(od))
         
         ### READ IN GWU DATA.  IF PICKLE FILE IS FOUND IN DIRECTORY, THIS IS USED FIRST.
@@ -112,12 +125,12 @@ if __name__=='__main__':
         picklefile=[afile for afile in infiles if afile.split('.')[-1]=='pickle']
         
         if len(picklefile)>1:
-            lf.write('Failed to load in folder, %s, had more than one .pickle files.\n'%folder)
+            lf.write('Failed to load in folder, %s, had more than one .pickle files.\n\n'%folder)
             raise IOError('Indirectory must contain only one pickle file if you want to use it!!!')
      
         elif len(picklefile)==1:
             df_full=df_load(picklefile[0])        
-            lf.write('Loaded contents of folder, %s, using the ".pickle" file, %s.\n'%(folder, get_shortname(picklefile[0])))            
+            lf.write('Loaded contents of folder, %s, using the ".pickle" file, %s.\n\n'%(folder, get_shortname(picklefile[0])))            
             
         ### If no picklefile found try reading in datafile directly
         else:
@@ -126,79 +139,127 @@ if __name__=='__main__':
                 if len(timefile) == 1:
                     infiles.remove(timefile[0])
                     df_full=from_timefile_datafile(datafile=infiles[0], timefile=timefile[0])
-                    lf.write('Loading contents of folder, %s, using the timefile/datafile conventional import\n'%folder)   
+                    lf.write('Loading contents of folder, %s, using the timefile/datafile conventional import\n\n'%folder)   
  
                 ### ACCOUNT FOR REMOTE POSSIBILITY THAT 2 FILES COULD STILL JUST BE TWO RAW DATA FILES INSTEAD OF DATAFILE/TIMEFILE
                 elif len(timefile) == 0:
                     df_full=from_spec_files(infiles)
                     lf.write('Loading contents of folder, %s, multiple raw spectral files %s.  (If these were \
                     supposed to be datafile/timefile and not raw files, couldnt find word "time" in filename.) \
-                    \n'%(folder, len(infiles))) 
+                    \n\n'%(folder, len(infiles))) 
 
                 else:
-                    lf.write('Failure: multiple timefile matches found in folder %s\n'%folder)                                
+                    lf.write('Failure: multiple timefile matches found in folder %s\n\n'%folder)                                
                     raise IOError('Timefile not found.  File must contain word "time"')                    
                 
             else:
                 df_full=from_spec_files(infiles)
-                lf.write('Loading contents of folder, %s, multiple raw spectral files %s.\n'%(folder, len(infiles)))            
-                
-                
-                
-
+                lf.write('Loading contents of folder, %s, multiple raw spectral files %s.\n\n'%(folder, len(infiles)))                    
+            
         ### Output the pickled dataframe    
         df_dump(df_full, outdir+'/rundata.pickle')
         
         ### Subtract the dark spectrum if it has one.  Note that all program should produce an attribute for darkseries,
         ### which may be None, but the attribute should still be here.
-        if hasattr(df_full, 'darkseries'):
+        if hasattr(df_full, 'darkseries') and sp.sub_base==True:
             if isinstance(df_full.darkseries, type(None) ):
                 df=df_full #need this                
-                lf.write('Warning: darkseries not found in data of run directory %s\n'%folder)            
+                lf.write('Warning: darkseries not found in data of run directory %s\n\n'%folder)            
             else:    
                 df=df_full.sub(df_full.darkseries, axis='index')
         else:
             df=df_full #need this
-            lf.write('Warning: darkseries attribute is no found on dataframe in folder %s!\n'%folder)            
+            lf.write('Warning: darkseries attribute is no found on dataframe in folder %s!\n\n'%folder)            
 
+        ### Fit first order baselines automatically?
+        if sp.line_fit:
+            blines=dynamic_baseline(df, sp.fit_regions )
+            df=df-blines 
+
+        ### Convert time axis to specified unit          
+        try:
+            timeunit=sp.timeunit #Take from params file  
+        except AttributeError:
+            timeunit='interval'
+        df.columns=datetime_convert(df.columns, return_as=timeunit)     
+       
+        ### Slice spectra and time start/end points.  Doesn't stop script upon erroringa
+        try:
+            df=df[sp.tstart: sp.tend] 
+        except Exception:
+            lf.write('Parameters Error: unable to slice tstart and tend to specified range in directory %s\n\n'%folder)           
         
-        ### Make parameter.  Really don't need this except autoscaling plots with the full data, primarily setting
-        ### the xlim manually and autoscaling the yaxis is a bit tricky.
-        df=df.ix[430.0: 680.0] 
-
-
-        ### Replace columns with default time unit (seconds)
-        timeunit='seconds' #Take from params file         
-        df.columns=datetime_convert(df.columns, return_as=timeunit)  
+        try:
+            df=df.ix[sp.x_min: sp.x_max] 
+        except Exception:
+            lf.write('Parameters Error: unable to slice x_min and x_max range in directory %s\n\n'%folder)      
+                 
+        
         transfer_attr(df_full, df, speakup=False)
+        df.columns.name=timeunit  #Useful for plotting 2d-3d for autodetection, but not necessary       
         
-
-        ### Fix colors after uploading    
-        specplot(df, title=options.rname+'full_spectrum', colors=None)
+        ###### Polygon Plot
+        spec_poly3d(df, title=options.rname+'3d Poly Spec')
+        plt_clrsave(outdir, options.rname+'polygon')
+        
+        ##### Basic spectral and absorbance plots    
+        specplot(df, title=options.rname+'Full spectrum')
         plt_clrsave(outdir, options.rname+'full_spectrum')
         
-        absplot(divby(df), title=options.rname+'relative_spectrum' )
+        absplot(divby(df), title=options.rname+'Relative spectrum' )
         plt_clrsave(outdir, options.rname+'relative')
        
-        ### Make parameter
-        uv_ranges=((430.0,450.0), (450.0,515.0),(515.0,570.0),(570.0, 620.0),(620.0, 680.0) )      
+        ### Look for uv-vis ranges in data, if not found, default to equally slicing spectrum by 7
+        try:
+            uv_ranges=sp.uv_ranges
+            if isinstance(uv_ranges, float) or isinstance(uv_ranges, int):
+                uv_ranges=spec_slice(df.index, uv_ranges)   
+                
+        except AttributeError:
+            uv_ranges=spec_slice(df.index, 8)   
        
-        ### Time averaged plot, not scaled to 1 (relative intenisty depend son bin width and actual intensity)
+        ### Time averaged plot, not scaled to 1 (relative intenisty dependson bin width and actual intensity)
         dfsliced=wavelength_slices(df, ranges=uv_ranges, apply_fcn='mean')
-        range_timeplot(dfsliced, ylabel='Average Intensity', xlabel='Time ('+timeunit+')' )
+        range_timeplot(dfsliced, ylabel='Average Intensity', xlabel='Time ('+timeunit+')' ) #legstyle =1 for upper left
         plt_clrsave(outdir, options.rname+'raw_time')
         
         ### Now scale curves to 1 for objective comparison
         dfsliced_norm=dfsliced.apply(lambda x: x/x[0], axis=1)
-        range_timeplot(dfsliced_norm, ylabel='Scaled Average Intensity', xlabel='Time ('+timeunit+')' )
+        range_timeplot(dfsliced_norm, title='Normalized Range Timeplot', ylabel='Scaled Average Intensity',\
+                       xlabel='Time ('+timeunit+')', legstyle=1 )
         plt_clrsave(outdir, options.rname+'norm_time')        
         
-        
+        ### Area plot using simpson method of integration
         dfarea=wavelength_slices(df, ranges=(min(df.index), max(df.index)), apply_fcn='simps')                
         range_timeplot(dfarea, ylabel='Power', xlabel='Time ('+timeunit+')', legend=False,
                        title='Spectral Power vs. Time (%i nm - %i nm)'%  ## Eventually make nm a paremeter and class method
-                       (min(df.index), max(df.index)), colors=None)
+                       (min(df.index), max(df.index)), color='r')
         plt_clrsave(outdir, options.rname+'full_area')
+        
+
+        #### correlation analysis
+        outcorr=od+'/corr_analysis'
+        os.mkdir(outcorr)        
+        ref=make_ref(df, method='empty')     
+
+        S,A=corr_analysis(df, ref)
+        sync_3d(A, title='Synchronous Spectrum (%s-%s %s)'%(round(min(df), 1), round(max(df),1), timeunit)) #min/max by columns      
+        plt_clrsave(outcorr, options.rname+'full_sync')                
+        async_3d(A, title='Asynchronous Spectrum (%s-%s %s)'%(round(min(df), 1), round(max(df),1), timeunit))
+        plt_clrsave(outcorr, options.rname+'full_async')        
+        
+        
+        absdf=divby(df) 
+        S,A=corr_analysis(absdf, ref)        
+        sync_3d(A, title='Synchronous Spectrum (%s-%s %s)'%(round(min(df), 1), round(max(df),1), timeunit)) #min/max by columns      
+        plt_clrsave(outcorr, options.rname+'relative_sync')                
+        async_3d(A, title='Asynchronous Spectrum (%s-%s %s)'%(round(min(df), 1), round(max(df),1), timeunit))
+        plt_clrsave(outcorr, options.rname+'relative_async')         
+        
+
+        #### 2d contour plot (solid color or cmap accepted)
+        plot2d(df, title='Full Contour', cmap='autumn', contours=7, label=1, colorbar=1, background=1)
+        plt_clrsave(outdir, options.rname+'full_contour')        
         
         
         
@@ -211,19 +272,19 @@ if __name__=='__main__':
             os.mkdir(out3d)
         
             for view in views:        
-                spec_surface3d(df, kind=kind, c_iso=c_iso, r_iso=r_iso, cmap=cm.autumn, elev=view[0], azim=view[1],
-                               xlabel='Time ('+timeunit+')')
+                spec_surface3d(df, kind=kind, c_iso=c_iso, r_iso=r_iso, cmap=cmget('gray'), contour_cmap=cmget('autumn'),
+                               elev=view[0], azim=view[1], xlabel='Time ('+timeunit+')')
                 plt_clrsave(out3d, '%sfull_3d_%s_%s'%(options.rname, view[0], view[1])   )    
                 
          
-                spec_surface3d(divby(df), kind=kind, c_iso=c_iso, r_iso=r_iso, cmap=cm.autumn, elev=view[0], azim=view[1],
-                               xlabel='Time ('+timeunit+')')
+                spec_surface3d(divby(df), kind=kind, c_iso=c_iso, r_iso=r_iso, cmap=cmget('gray'), contour_cmap=cmget('autumn'),
+                               elev=view[0], azim=view[1], xlabel='Time ('+timeunit+')')
                 plt_clrsave(out3d, '%srelative_3d_%s_%s'%(options.rname, view[0], view[1])   )    
         
 
         
 
     ### Close logfile    
-    lf.write('Time for completion, %s'%(round(time.time()-start),1))
+    lf.write('Time for completion, %s'%(round(time.time()-start,1)))
     lf.close()