+++ Changes.rst
@@ -1,1 +1,5 @@
 Version 0.1.0, 11/16/12 -- Initial Internal Release
+
+Version 0.1.1, 12/1/12 -- Updated several tools, especially 2d-3d plots and color mapping.l
+                          Created GWU curve analysis script, gwu_spec.py
+			  Added 2d correlation analysis.